requirements.txt
pandas==2.1.4
matplotlib==3.8.2
tkinter
pillow==10.2.0
plotly==5.18.0 